Changed gRPC protocol to thrift by default
